+++ gdpr/src/main/java/com/kairos/controller/data_inventory/processing_activity/OrganizationProcessingPurposeController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
-import java.math.BigInteger;
 
 import static com.kairos.constants.ApiConstant.API_ORGANIZATION_UNIT_URL;
 import static com.kairos.constants.ApiConstant.COUNTRY_URL;
@@ -46,7 +45,7 @@ public class OrganizationProcessingPurposeController {
 
     @ApiOperation("get processing purpose by id")
     @GetMapping("/processing_purpose/{processingPurposeId}")
-    public ResponseEntity<Object> getProcessingPurpose(@PathVariable Long unitId, @PathVariable BigInteger processingPurposeId) {
+    public ResponseEntity<Object> getProcessingPurpose(@PathVariable Long unitId, @PathVariable Long processingPurposeId) {
         if (unitId == null) {
             return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
         }
@@ -66,7 +65,7 @@ public class OrganizationProcessingPurposeController {
 
     @ApiOperation("delete processing purpose by id")
     @DeleteMapping("/processing_purpose/{processingPurposeId}")
-    public ResponseEntity<Object> deleteProcessingPurpose(@PathVariable Long unitId, @PathVariable BigInteger processingPurposeId) {
+    public ResponseEntity<Object> deleteProcessingPurpose(@PathVariable Long unitId, @PathVariable Long processingPurposeId) {
         if (unitId == null) {
             return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
         }
@@ -76,7 +75,7 @@ public class OrganizationProcessingPurposeController {
 
     @ApiOperation("update processing purpose by id")
     @PutMapping("/processing_purpose/{processingPurposeId}")
-    public ResponseEntity<Object> updateProcessingPurpose(@PathVariable Long unitId, @PathVariable BigInteger processingPurposeId, @Valid @RequestBody ProcessingPurposeDTO processingPurpose) {
+    public ResponseEntity<Object> updateProcessingPurpose(@PathVariable Long unitId, @PathVariable Long processingPurposeId, @Valid @RequestBody ProcessingPurposeDTO processingPurpose) {
         if (unitId == null) {
             return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
         }
